Move extension.py to configuration_v1
